requirements.txt
sqlalchemy
psycopg2-binary
Flask
openai>=1.0.0
prisma-client-py
